+++ doc/source/history.rst
@@ -6,7 +6,7 @@ Release history
 0.8.0 (current development version)
 ===================================
 
-Under construction.
+* CHANGED: Updated test suite to support ``pytest`` 5.0+.
 
 
 0.7.4 (17 May 2019)